Functions to add a neuron.
New functions allow a neuron to be added to a network, even while the
network is training.  The main new function is add-connected-neuron.
Also created some other helpful functions for navigating and updating
the neural network.  Here's the full list of new functions.

  * collect-neurons
  * reset-neuron-weights
  * add-connected neuron
  * add-incoming-cxs (not idempotent!)
  * add-outgoing-cxs (not idempotent!)
  * get-neuron-by-name

Added target-error parameter to test-train, and made it so that if you
provide a value for weights-file, then weights are not reset,
regardless of the value of reset-weights.