Refined training functions.
Training now takes the number of available CPUs into account when
determining the default number of threads to use for training.

Added support for multiple neural-network and training environments.
An environment consists of a neural network, a training file, and
a test file, and is identified by a keyword ID.

Added new training functions:
  * create-environment
  * remove-environment
  * list-environments
  * train
  * stop-training

Removed the old training functions.

Included 2 sets of example training and testing sets, all in one
compressed file, and configured git-lfs to properly manage the
compressed file.