Added function training-set->pngs.
This function iterates through the training set, creating
PNG files in the file system, with each training set vector
being converted into a PNG that is stored in folder named
according to the label that represents the vector's expected
output classification.

For example, if the training-set vectors represent images
of digits, then the training-set->pngs function creates
the images and stores them in a folder corresponding to
the digit.