Sped up training by a factor of 5.
Changed the training mechanism so that instead of calculating network
error independently at every iteration, I instead keep a vector with
training-frame errors, focus the training on the frames where the
error is above the target error, and average the errors in the
vector to determine network error.  Because the algorithm is able to
skip over the frames with low error, the computer can focus all of its
processing power on the frames that need it, and training speed
increases dramatically.

Consider training on the digits data (mnist-train.csv) on host
bianca-1, using 34 hidden neurons and 8 threads. Whereas the best time
for that was about 1 hour before this commit, now the time is under 12
minutes.

This is, nevertheless, much, much slower than something like TensorFlow, Torch,
Theano, or Julia's Flux, so there's still a lot of work to do.  (Still
need to get comparable timings on those other systems, starting with
Flux, but I'm guessing at least one order of magnitude.)