Added the ability to connect mostly.
Connect mostly allows every n'th connection to be skipped when
creating connections.  That way, a given neuron is not connected to
all the neurons in the next layer, but rather most of the nueruons in
the next layer.

Removed train-bad-frame-1 and related code because it proved not to
help with reducing training time or increasing accuracy.

Trained with a bunch of different configurations and recorded the
results in notes.org.